+++ lib/src/features/main/my_card/my_card_vm.dart
@@ -1,0 +1,9 @@
+import 'package:mobx/mobx.dart';
+
+part 'my_card_vm.g.dart';
+
+class MyCardVm = _MyCardVm with _$MyCardVm;
+
+abstract class _MyCardVm with Store {
+  _MyCardVm();
+}