+++ lib/src/features/onboard/ui/onboarding_vm.dart
@@ -1,11 +1,9 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:mobx/mobx.dart';
 import 'package:routemaster/routemaster.dart';
 import 'package:where_to_go_today/src/core/ui/base/view_model.dart';
 import 'package:where_to_go_today/src/core/ui/errors_handling/error_handler.dart';
-import 'package:where_to_go_today/src/features/main/main_screen_route.dart';
+import 'package:where_to_go_today/src/features/main/main_route.dart';
 import 'package:where_to_go_today/src/features/onboard/services/bloc/events/onboarding_event.dart';
 import 'package:where_to_go_today/src/features/onboard/services/bloc/states/onboarding_state.dart';
 import 'package:where_to_go_today/src/features/onboard/services/onboarding_bloc.dart';
@@ -49,6 +47,7 @@ abstract class _OnboardingVm extends ViewModel with Store {
 
   void onSkipButtonClick() {
     _bloc.add(const OnboardingEventSkipped());
+    Routemaster.of(_context).push(MainRoute.routeName);
   }
 
   @action
@@ -58,7 +57,7 @@ abstract class _OnboardingVm extends ViewModel with Store {
 
   void _handleStates(OnboardingState state) {
     if (state is OnboardingStateSkip) {
-      Routemaster.of(_context).push(MainScreenRoute.routeName);
+      Routemaster.of(_context).push(MainRoute.routeName);
     }
   }
 }